requirements.txt
altair==6.0.0
annotated-types==0.7.0
ansible-core==2.18.12
argcomplete==3.6.3
asgiref==3.11.0
attrs==25.4.0
bcc==0.35.0
beautifulsoup4==4.14.3
blinker==1.9.0
blivet==3.12.1
blivet-gui==2.6.0
boto3==1.42.15
botocore==1.42.15
Brlapi==0.8.7
cachetools==6.2.4
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.1
cockpit @ file:///builddir/build/BUILD/cockpit-353.1-build/cockpit-353.1/tmp/wheel/cockpit-353.1-py3-none-any.whl#sha256=7bcbb10e8083042579647d9004814730b341e2ca7228ea7e8ebd6fceb7f5d870
contourpy==1.3.3
crypt_r==3.13.1
cryptography==45.0.4
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-python==1.4.0
dbus_next==0.2.3
distro==1.9.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
dnf==4.24.0
dnsconfd==1.7.3
dnspython==2.8.0
file-magic==0.4.0
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.45
icoextract==0.1.5
idna==3.10
iniconfig==2.3.0
iso639==0.1.4
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.9
langtable==0.0.69
libcomps==0.1.23
libdnf==0.75.0
libdnf5==5.2.17.0
louis==3.33.0
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib==3.10.8
mplsoccer==1.6.1
narwhals==2.14.0
netaddr==1.3.0
nftables==0.1
numpy==2.4.0
olefile==0.47
packaging==25.0
pandas==2.3.3
pefile==2024.8.26
perf==0.1
pexpect==4.9.0
pid==3.0.4
pillow==11.3.0
plotly==6.5.0
pluggy==1.6.0
ply==3.11
productmd==1.49
protobuf==6.33.2
ptyprocess==0.7.0
pwquality==1.4.5
pyarrow==22.0.0
pycairo==1.28.0
pycparser==2.22
pycups==2.0.4
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyGObject==3.54.5
pyinotify==0.9.6
pykickstart @ file:///builddir/build/BUILD/pykickstart-3.66-build/pykickstart-3.66
pymongo==4.15.5
pyparsing==3.3.1
pyparted==3.13.0
PySocks==1.7.1
pytest==9.0.2
python-augeas==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
pytz==2025.2
pyudev==0.24.3
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.17.12-build/kernel-6.17.12/linux-6.17.12-300.fc43.x86_64/tools/net/ynl
referencing==0.36.2
requests==2.32.5
requests-file==2.0.0
requests-ftp==0.3.1
resolvelib==1.0.1
rpds-py==0.27.0
rpm==6.0.0
s3transfer==0.16.0
scipy==1.16.3
seaborn==0.13.2
selinux @ file:///builddir/build/BUILD/libselinux-3.9-build/libselinux-3.9/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.9-build/selinux-3.9/python/sepolicy
setools==4.6.0
setuptools==78.1.1
simpleline==1.9.0
six==1.17.0
smmap==5.0.2
sos==4.10.2
soupsieve==2.8.1
sqlparse==0.5.5
streamlit==1.52.2
systemd-python==235
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
umstellar==0.2.0
urllib3==2.6.2
watchdog==6.0.0
xkbregistry==0.3
